refactor(download): extract bitnet-download-core microcrate
